Extracted accounts ui out of dashboard folder and moved into dashboard-account
