+++ src/app/lib/ui/dashboard-account/account-card.jsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { UserIcon, WrenchIcon } from "@heroicons/react/24/outline";
 import { ArrowRightCircleIcon } from "@heroicons/react/24/solid";
-import { formatCurrency } from "../../../utils";
+import { formatCurrency } from "../../utils";
 
 const trackingType = {
   manual: { icon: UserIcon, text: "MANUAL TRACKING" },
